requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
google-generativeai==0.8.3
